requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
Django==1.11.4
djangorestframework==3.6.3
idna==2.6
pycparser==2.18
pyOpenSSL==17.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
